Separate fomo app domain from stock clients and save single quote data
Previously our models for querying stock data and saving single quote data was tied to alpha vantage.

Now the models are based on our own domain and translate to alphavantage when hitting the infrastructure layer to call alpha vantage. Our own domain objects are saved to queue and cache instead as well.

The new classes are StockQuery -> SingleQuoteQuery.

This required rewriting a lot of the classes and flow to save the single quote data when executing the queries from the queue. A QueryContext class was created to direct where the should be saved for each query type. In the future we may have many types.