Add chat msg status show if it's been sent or not and fix scrolling
When a chat message has been sent it will start as pending status and opacity 0.6. When the message is postbacked from server it will be changed to sent and opacity 1. Error will show an error icon.

Scrolling will now scroll to bottom of the chatbox and not scroll the whole window.